requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
behave==1.2.6
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.18.0
google-auth==2.16.0
google-auth-oauthlib==0.7.1
gspread==5.11.3
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mycdp==1.1.1
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.2
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pdbp==1.7.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pynose==1.5.4
pyotp==2.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
sbvirtualdisplay==1.4.0
selenium==4.30.0
seleniumbase==4.36.4
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabcompleter==1.4.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
webdriver-manager==3.8.6
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
